[frontend]: replace alpine with distroless image; build with only prod dependencies
