+++ src/frontend/components/CurrencySwitcher/CurrencySwitcher.tsx
@@ -5,7 +5,7 @@ import { useMemo } from 'react';
 import getSymbolFromCurrency from 'currency-symbol-map';
 import { useCurrency } from '../../providers/Currency.provider';
 import * as S from './CurrencySwitcher.styled';
-import { CypressFields } from '../../utils/Cypress';
+import { CypressFields } from '../../utils/enums/CypressFields';
 
 const CurrencySwitcher = () => {
   const { currencyCodeList, setSelectedCurrency, selectedCurrency } = useCurrency();
