+++ src/frontend/components/Recommendations/Recommendations.tsx
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-import { CypressFields } from '../../utils/Cypress';
+import { CypressFields } from '../../utils/enums/CypressFields';
 import { useAd } from '../../providers/Ad.provider';
 import ProductCard from '../ProductCard';
 import * as S from './Recommendations.styled';
