+++ src/frontend/pages/index.tsx
@@ -10,7 +10,7 @@ import * as S from '../styles/Home.styled';
 import { useQuery } from '@tanstack/react-query';
 import ApiGateway from '../gateways/Api.gateway';
 import Banner from '../components/Banner';
-import { CypressFields } from '../utils/Cypress';
+import { CypressFields } from '../utils/enums/CypressFields';
 import { useCurrency } from '../providers/Currency.provider';
 
 const Home: NextPage = () => {
